# GreenGearAPI/public/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "GreenGear API",
		"version": "1.0.0",
		"contact": {}
	},
	"servers": [
		{
			"url": "http://localhost"
		}
	],
	"paths": {
...
		"/api/users/{user_id}": {
			"get": {
				"tags": [
					"Users"
				],
				"summary": "Get User By ID",
				"description": "Get User By ID",
				"operationId": "getUserById",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"put": {
				"tags": [
					"Users"
				],
				"summary": "Update User",
				"description": "Update User",
				"operationId": "updateUser",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"example": "testuser@example.comupdate"
									},
									"first_name": {
										"type": "string",
										"example": "Testupdate"
									},
									"last_name": {
										"type": "string",
										"example": "Userupdate"
									},
									"password": {
										"type": "string",
										"example": "password123update"
									},
									"role": {
										"type": "string",
										"example": "user"
									},
									"username": {
										"type": "string",
										"example": "testuser update"
									}
								}
							},
							"example": {
								"email": "testuser@example.comupdate",
								"first_name": "Testupdate",
								"last_name": "Userupdate",
								"password": "password123update",
								"role": "user",
								"username": "testuser update"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"delete": {
				"tags": [
					"Users"
				],
				"summary": "Delete User",
				"description": "Delete User",
				"operationId": "deleteUser",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "user_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": ""
					}
				}
			]
		},
		"/api/login": {
			"post": {
				"tags": [
					"Users"
				],
				"summary": "Login",
				"description": "Login",
				"operationId": "login",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"password": {
										"type": "string",
										"example": "password123"
									},
									"username": {
										"type": "string",
										"example": "testuser"
									}
								}
							},
							"example": {
								"password": "password123",
								"username": "testuser"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		},
...
		"/api/gardens/{garden_id}": {
			"get": {
				"tags": [
					"Gardens"
				],
				"summary": "Get Garden By ID",
				"description": "Get Garden By ID",
				"operationId": "getGardenById",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"put": {
				"tags": [
					"Gardens"
				],
				"summary": "Update Garden",
				"description": "Update Garden",
				"operationId": "updateGarden",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"garden_name": {
										"type": "string",
										"example": "My Garden2"
									},
									"userId": {
										"type": "number",
										"example": 2
									}
								}
							},
							"example": {
								"garden_name": "My Garden2",
								"userId": 2
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"delete": {
				"tags": [
					"Gardens"
				],
				"summary": "Delete Garden",
				"description": "Delete Garden",
				"operationId": "deleteGarden",
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "garden_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": ""
					}
				}
			]
		},
...
		"/api/hub/": {
			"post": {
				"tags": [
					"Hub"
				],
				"summary": "Create Hub",
				"description": "Create Hub",
				"operationId": "createHub",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"garden_id": {
										"type": "number",
										"example": 1
									},
									"hub_name": {
										"type": "string",
										"example": "Hub#2"
									},
									"hub_token": {
										"type": "string",
										"example": "test_token"
									},
									"hub_version": {
										"type": "string",
										"example": "0.0.1"
									}
								}
							},
							"example": {
								"garden_id": 1,
								"hub_name": "Hub#2",
								"hub_token": "test_token",
								"hub_version": "0.0.1"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			}
		},
...
		"/api/hub/{hub_id}": {
			"get": {
				"tags": [
					"Hub"
				],
				"summary": "Get Hub by ID",
				"description": "Get Hub by ID",
				"operationId": "getHubById",
				"requestBody": {
					"content": {
						"text/plain": {
							"example": ""
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"put": {
				"tags": [
					"Hub"
				],
				"summary": "Update Hub",
				"description": "Update Hub",
				"operationId": "updateHub",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"hub_name": {
										"type": "string",
										"example": "Hub#2qwe"
									},
									"hub_token": {
										"type": "string",
										"example": "test_token"
									},
									"hub_version": {
										"type": "string",
										"example": "0.0.1"
									}
								}
							},
							"example": {
								"hub_name": "Hub#2qwe",
								"hub_token": "test_token",
								"hub_version": "0.0.1"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"delete": {
				"tags": [
					"Hub"
				],
				"summary": "Delete Hub",
				"description": "Delete Hub",
				"operationId": "deleteHub",
				"requestBody": {
					"content": {
						"text/plain": {
							"example": ""
						}
					}
				},
				"responses": {
					"200": {
						"description": ""
					}
				}
			},
			"parameters": [
				{
					"name": "hub_id",
					"in": "path",
					"required": true,
					"schema": {
						"type": "string",
						"example": ""
					}
				}
			]
		},
...
	},
...
}